+++ src/imgtools/coretypes/masktypes/__init__.py
@@ -1,0 +1,15 @@
+from .roi_matching import (
+    ROI_HANDLING,
+    ROIMatcher,
+    ROIMatchResult,
+    handle_roi_matching,
+    match_roi,
+)
+
+__all__ = [
+    "ROI_HANDLING",
+    "ROIMatcher",
+    "ROIMatchResult",
+    "handle_roi_matching",
+    "match_roi",
+]
